FINISHED TASK including lemon and spicy

+++ api/urls/urls.controllers.js
@@ -6,14 +6,14 @@ const baseUrl = "http:localhost:8090";
 
 exports.shorten = async (req, res) => {
   // create url code
+  console.log("I am in shorten function");
   const urlCode = shortid.generate();
-  console.log(req.user);
   try {
     req.body.shortUrl = baseUrl + "/" + urlCode;
     req.body.urlCode = urlCode;
-    req.body.userId = req.params.userId;
+    req.body.userId = req.user._id;
     const newUrl = await Url.create(req.body);
-    await User.findByIdAndUpdate(req.params.userId, {
+    await User.findByIdAndUpdate(req.user._id, {
       $push: { urls: newUrl._id },
     });
     res.json(newUrl);
@@ -38,6 +38,9 @@ exports.redirect = async (req, res) => {
 exports.deleteUrl = async (req, res) => {
   try {
     const url = await Url.findOne({ urlCode: req.params.code });
+    if (req.user._id.toString() !== url.userId.toString()) {
+      return res.status(401).json("You can only delete your own URLS");
+    }
     if (url) {
       await Url.findByIdAndDelete(url._id);
       return res.status(201).json("Deleted");
